requirements.txt
azure-ai-textanalytics==5.3.0
azure-ai-language-questionanswering # for QnA Solutions
python-dotenv